requirements.txt
blinker==1.9.0
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
pillow==11.2.1
python-dateutil==2.9.0.post0
qrcode==8.2
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
urllib3==2.4.0
waitress==3.0.2
Werkzeug==3.1.3
